+++ src/utils/constants.ts
@@ -1,0 +1,2 @@
+
+export const defaultConfigName = ".lr.module.gen"