+++ src/utils/prompt/index.ts
@@ -1,0 +1,9 @@
+import { askModule } from "./askModulePrompt"
+import { askVariables } from "./askVariablesPrompt"
+
+const Prompt = {
+    askModule,
+    askVariables
+}
+
+export default Prompt