+++ src/utils/typeCheckers.ts
@@ -1,0 +1,1 @@
+export const isStringArray = (type: string[] | any): type is string[] => Array.isArray(type)